daemon: rework storage architecture
