+++ xelis_daemon/src/core/storage/providers/mod.rs
@@ -1,0 +1,21 @@
+mod asset;
+mod blocks_at_height;
+mod dag_order;
+mod difficulty;
+mod pruned_topoheight;
+mod nonce;
+mod balance;
+mod client_protocol;
+mod transaction;
+mod block;
+
+pub use asset::AssetProvider;
+pub use blocks_at_height::BlocksAtHeightProvider;
+pub use dag_order::DagOrderProvider;
+pub use difficulty::DifficultyProvider;
+pub use pruned_topoheight::PrunedTopoheightProvider;
+pub use nonce::NonceProvider;
+pub use balance::BalanceProvider;
+pub use client_protocol::ClientProtocolProvider;
+pub use transaction::TransactionProvider;
+pub use block::BlockProvider;
